[EDIT] delete example, make firebase config more neatly
