--- _firebase_Config.ts
@@ -1,12 +1,0 @@
-const firebaseConfig = {
-  apiKey: 'AIzaSyCQWLkImcBE1BlCM525V28xOZ1YAIP0LfA',
-  authDomain: 'moinda-2c664.firebaseapp.com',
-  databaseURL: 'https://moinda-2c664.firebaseio.com',
-  projectId: 'moinda-2c664',
-  storageBucket: 'moinda-2c664.appspot.com',
-  messagingSenderId: '984280015557',
-  appId: '1:984280015557:web:6da6e95c999f5cfb7486ac',
-  measurementId: 'G-MCRTQB8SVN',
-};
-
-export { firebaseConfig };
